Ajout de la brique de base du jeu et amélioration des objets JSON

+++ src/main/java/fr/la7prod/maze/util/Direction.java
@@ -19,10 +19,10 @@ public enum Direction {
 	}
 
 	public int getX() {
-		return c.getX();
+		return (int)c.getX();
 	}
 	public int getY() {
-		return c.getY();
+		return (int)c.getY();
 	}
 
 	public Direction getOpposite() {
